+++ scripts/consts.js
@@ -1,0 +1,12 @@
+export const MODULE_ID = 'roll-from-compendium'
+export const MODULE_NAME = 'Quick Roll To Chat'
+export const COMPATIBLE_DOCUMENT_TYPES = [
+  'JournalEntry',
+  'Actor',
+  'Scene',
+  'Macro',
+  'RollTable',
+  'Item',
+]
+export const DUMMY_ACTOR_IMAGE = 'icons/svg/d20-highlight.svg'
+export const DUMMY_ACTOR_NAME = '(Quick Roll To Chat)'