+++ scripts/pf2e-compatibility.js
@@ -1,7 +1,7 @@
 // https://gitlab.com/hooking/foundry-vtt---pathfinder-2e/-/blob/master/src/module/actor/sheet/base.ts#L1048
 export const pf2eInitializeDummyActor = async (compendiumRollActor) => {
-  // setting to Level -2, for total Trained modifier of +0
-  await compendiumRollActor.update({ data: { details: { level: { value: -2 } } } })
+  // setting to Level 0, for total Trained modifier of +2.  (-2 is no longer possible)
+  await compendiumRollActor.update({ data: { details: { level: { value: 0 } } } })
   await createSpellcastingEntry(compendiumRollActor)
   return compendiumRollActor
 }
@@ -12,7 +12,7 @@ const createSpellcastingEntry = (compendiumRollActor) => {
   const ability = 'int'
   const flexible = false
 
-  const name = '(Compendium Roll Spellcasting)'
+  const name = '(Quick Roll To Chat; Spellcasting)'
 
   // Define new spellcasting entry
   const spellcastingEntry = {
@@ -35,34 +35,52 @@ const createSpellcastingEntry = (compendiumRollActor) => {
 const getSpellcasting = (actor, dummyActor) => {
   const existingSpellcasting = actor.spellcasting.filter(sc => sc)[0]
   if (existingSpellcasting) return existingSpellcasting
-  else return dummyActor.spellcasting.getName('(Compendium Roll Spellcasting)')
+  else return dummyActor.spellcasting.getName('(Quick Roll To Chat; Spellcasting)')
 }
 
-export const pf2eCastSpell = (item, actor, dummyActor) => {
+export const pf2eCastSpell = async (item, actor, dummyActor) => {
   const spellcasting = getSpellcasting(actor, dummyActor)
   Object.defineProperty(item, 'spellcasting', {
     value: spellcasting,
     configurable: true,
   })
-  return item.toMessage(undefined, { create: false }).then(async chatMessage => {
-    const dataItemId = `data-item-id="${item.id}"`
-    item.data.data.location.value = spellcasting.id
-    const dataEmbeddedItem = `data-embedded-item="${escapeHtml(JSON.stringify(item.toObject(false)))}"`
-    const editedContent = chatMessage.data.content.replace(dataItemId, `${dataItemId} ${dataEmbeddedItem}`)
-    await chatMessage.data.update({ content: editedContent })
-    return ChatMessage.create(chatMessage.data)
-  })
+  const shiftPressed = game.keyboard.isModifierActive(KeyboardManager.MODIFIER_KEYS.SHIFT)
+  const overrideSpellLevel = shiftPressed ? await upcastSpellLevel(item) : undefined
+
+  const originalAutoHeightenLevel = item.data.data.location.autoHeightenLevel
+  item.data.data.location.autoHeightenLevel = overrideSpellLevel || originalAutoHeightenLevel
+  const chatMessage = await item.toMessage(undefined, { create: false, data: { spellLvl: overrideSpellLevel } })
+
+  const dataItemId = `data-item-id="${item.id}"`
+  item.data.data.location.value = spellcasting.id
+  const dataEmbeddedItem = `data-embedded-item="${escapeHtml(JSON.stringify(item.toObject(false)))}"`
+  const editedContent = chatMessage.data.content.replace(dataItemId, `${dataItemId} ${dataEmbeddedItem}`)
+  await chatMessage.data.update({ content: editedContent })
+
+  return ChatMessage.create(chatMessage.data)
 }
 
-export const pf2eItemToMessage = (item) => {
+export const pf2eItemToMessage = async (item) => {
   const originalItemDataType = item.data.type
   if (['ancestry', 'background', 'class'].includes(originalItemDataType)) {
     item.data.type = 'feat'
   }
-  return item.toMessage().then(async chatMessage => {
-    item.data.type = originalItemDataType
-    return chatMessage
-  })
+  const chatMessage = await item.toMessage()
+  item.data.type = originalItemDataType
+
+  if (['effect', 'condition'].includes(originalItemDataType)) {
+    // add an extra chat message for draggable effects
+    const contentStr = item.pack
+      ? `@Compendium[${item.pack}.${item.id}]{${item.name}}`
+      : `@Item[${item.id}]{${item.name}}`
+    await ChatMessage.create({
+      user: game.user.id,
+      speaker: { user: game.user, alias: `Draggable ${originalItemDataType}:` },
+      content: contentStr,
+    })
+  }
+
+  return chatMessage
 }
 
 function escapeHtml (string) {
@@ -79,4 +97,56 @@ function escapeHtml (string) {
   return String(string).replace(/[&<>"'`=\/]/g, function (s) {
     return replacements[s]
   })
+}
+
+const upcastSpellLevel = async (item) => {
+  let content = `
+<div>
+    <div class="form-group">
+        <select id="selectedLevel">`
+  for (const spellLevel of [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]) {
+    if (spellLevel === item.level) {
+      content += `<option value="${spellLevel}" selected>${spellLevel}${th(spellLevel)} Level (Base)</option>`
+    } else if (spellLevel > item.level) {
+      content += `<option value="${spellLevel}">${spellLevel}${th(spellLevel)} Level (+${spellLevel - item.level})</option>`
+    }
+  }
+  content += `</select>
+    </div>
+</div>`
+
+  return new Promise((resolve, reject) => {
+    new Dialog({
+      title: `Upcast ${item.name}`,
+      content: content,
+      buttons: {
+        cast: {
+          label: 'Cast',
+          callback: html => {
+            const spellLevel = parseInt(html.find('#selectedLevel')[0].value)
+            resolve(spellLevel)
+          }
+        },
+        cancel: {
+          label: 'Cancel',
+          callback: () => {
+            reject()
+          }
+        }
+      },
+      default: 'cancel',
+    }).render(true)
+  })
+}
+
+const th = (num) => {
+  if (num % 10 === 1 && num % 100 !== 11) {
+    return 'st'
+  } else if (num % 10 === 2 && num % 100 !== 12) {
+    return 'nd'
+  } else if (num % 10 === 3 && num % 100 !== 13) {
+    return 'rd'
+  } else {
+    return 'th'
+  }
 }